Remove Phoenix namespace; move models to Models.swift
